Add user authentication in frontend

+++ backend/src/controllers/todos.ts
@@ -3,10 +3,15 @@ import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
 
 import TodoModel from '../models/todo';
+import { assertIsDefined } from '../util/assertIsDefined';
 
 export const getTodos: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   try {
-    const todos = await TodoModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const todos = await TodoModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(todos);
   } catch (error) {
     next(error);
@@ -15,8 +20,11 @@ export const getTodos: RequestHandler = async (req, res, next) => {
 
 export const getTodo: RequestHandler = async (req, res, next) => {
   const todoId = req.params.todoId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(todoId)) {
       throw createHttpError(400, 'Invalid todo id');
     }
@@ -25,6 +33,10 @@ export const getTodo: RequestHandler = async (req, res, next) => {
 
     if (!todo) {
       throw createHttpError(404, 'Todo not found');
+    }
+
+    if (!todo.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'Unauthorized request');
     }
 
     res.status(200).json(todo);
@@ -46,13 +58,17 @@ export const postTodo: RequestHandler<
 > = async (req, res, next) => {
   const text = req.body.text;
   const isCompleted = req.body.isCompleted;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!text) {
       throw createHttpError(400, 'Text is required');
     }
 
     const newTodo = await TodoModel.create({
+      userId: authenticatedUserId,
       text: text,
       isCompleted: isCompleted,
     });
@@ -71,12 +87,20 @@ interface UpdateTodoBody {
   isCompleted?: boolean;
 }
 
-export const updateTodo: RequestHandler<UpdateTodoParams, unknown, UpdateTodoBody, unknown> = async (req, res, next) => {
+export const updateTodo: RequestHandler<
+  UpdateTodoParams,
+  unknown,
+  UpdateTodoBody,
+  unknown
+> = async (req, res, next) => {
   const todoId = req.params.todoId;
   const newText = req.body.text;
   const newIsCompleted = req.body.isCompleted;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(todoId)) {
       throw createHttpError(400, 'Invalid todo id');
     }
@@ -95,6 +119,10 @@ export const updateTodo: RequestHandler<UpdateTodoParams, unknown, UpdateTodoBod
       throw createHttpError(404, 'Todo not found');
     }
 
+    if (!todo.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'Unauthorized request');
+    }
+
     todo.text = newText;
     todo.isCompleted = newIsCompleted;
 
@@ -107,8 +135,11 @@ export const updateTodo: RequestHandler<UpdateTodoParams, unknown, UpdateTodoBod
 
 export const deleteTodo: RequestHandler = async (req, res, next) => {
   const todoId = req.params.todoId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(todoId)) {
       throw createHttpError(400, 'Invalid todo id');
     }
@@ -119,9 +150,13 @@ export const deleteTodo: RequestHandler = async (req, res, next) => {
       throw createHttpError(404, 'Todo not found');
     }
 
+    if (!todo.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'Unauthorized request');
+    }
+
     await todo.deleteOne();
     res.sendStatus(204);
   } catch (error) {
     next(error);
   }
-}
+};
